requirements/production.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile production.in
#
amqp==5.0.6
    # via kombu
arrow==1.1.1
    # via -r production.in
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
backcall==0.2.0
    # via ipython
billiard==3.6.4.0
    # via celery
celery[redis]==5.1.2
    # via
    #   -r production.in
    #   django-celery-beat
    #   django-post-request-task
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.0.3
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
decorator==5.0.9
    # via ipython
django==3.2.7
    # via
    #   -r production.in
    #   django-appconf
    #   django-celery-beat
    #   django-extensions
    #   django-filter
    #   django-post-request-task
    #   django-sql-utils
    #   django-timezone-field
    #   djangorestframework
django-appconf==1.0.4
    # via django-imagekit
django-celery-beat==2.2.1
    # via -r production.in
django-extensions==3.1.3
    # via -r production.in
django-filter==2.4.0
    # via -r production.in
git+https://github.com/matthewwithanm/django-imagekit.git@01358ca83d3876ac77425de37b771070fd429c82#egg=django-imagekit
    # via -r production.in
django-object-actions==3.0.2
    # via -r production.in
django-post-request-task==0.5
    # via -r production.in
django-sql-utils==0.6.1
    # via -r production.in
django-timezone-field==4.2.1
    # via django-celery-beat
djangorestframework==3.12.4
    # via -r production.in
factory-boy==3.2.0
    # via -r production.in
faker==8.12.1
    # via factory-boy
iniconfig==1.1.1
    # via pytest
ipython==7.27.0
    # via -r production.in
jedi==0.18.0
    # via ipython
kombu==5.1.0
    # via celery
matplotlib-inline==0.1.2
    # via ipython
packaging==21.0
    # via pytest
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
pilkit==2.0
    # via django-imagekit
pillow==8.3.1
    # via -r production.in
pluggy==1.0.0
    # via pytest
prompt-toolkit==3.0.20
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.1
    # via -r production.in
ptyprocess==0.7.0
    # via pexpect
py==1.10.0
    # via pytest
pygments==2.10.0
    # via ipython
pyparsing==2.4.7
    # via packaging
pytest==6.2.5
    # via
    #   pytest-django
    #   pytest-mock
pytest-django==4.4.0
    # via -r production.in
pytest-mock==3.6.1
    # via -r production.in
python-crontab==2.5.1
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   arrow
    #   faker
    #   python-crontab
pytz==2021.1
    # via
    #   celery
    #   django
    #   django-timezone-field
redis==3.5.3
    # via celery
six==1.16.0
    # via
    #   click-repl
    #   django-imagekit
    #   python-dateutil
sqlparse==0.4.1
    # via
    #   django
    #   django-sql-utils
text-unidecode==1.3
    # via faker
toml==0.10.2
    # via pytest
traitlets==5.1.0
    # via
    #   ipython
    #   matplotlib-inline
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
